requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
cssselect==1.0.3
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-js-asset==1.1.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-rest-framework-mongoengine==3.3.1
django-reversion==3.0.2
django-simple-captcha==0.5.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.0
mongoengine==0.16.3
mysqlclient==1.3.13
odfpy==1.4.0
openpyxl==2.5.12
parsel==1.5.1
Pillow==5.3.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pymongo==3.7.2
PyMySQL==0.9.3
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.7
pywin32==224
PyYAML==3.13
queuelib==1.5.0
redis==3.0.1
requests==2.21.0
Scrapy==1.5.1
scrapy-redis==0.6.8
service-identity==18.1.0
six==1.12.0
tablib==0.12.1
Twisted==18.9.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
w3lib==1.19.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
